+++ CHANGELOG.rst
@@ -3,6 +3,22 @@ Changelog
 ==========
 
 +++++++++
+v0.1.0 (19/04/2024)
++++++++++
+
+**Added**
+
+- Extra detail to error messages
+- Filter for invalid names
+
+
+**Fixed**
+
+- Unassigned item now updates properly once saved
+- Edit button is no longer available if shipment has been booked
+- Samples table now redirects to correct sample
+
++++++++++
 v0.0.1 (27/03/2024)
 +++++++++